chore: some refactoring
