+++ src/renderer/features/operations/OperationsConfirm/RemovePureProxied/model/confirm-model.ts
@@ -1,28 +1,24 @@
 import { createEvent, combine, restore } from 'effector';
 
-import type { Chain, Address, ProxiedAccount, Account, Transaction } from '@shared/core';
-import { ProxyType } from '@shared/core';
+import { Account, AccountId, Chain, ProxiedAccount, ProxyType, Transaction } from '@shared/core';
 import { networkModel } from '@entities/network';
 import { walletModel, walletUtils } from '@entities/wallet';
 
 type Input = {
-  chain: Chain;
-  account: Account;
   signatory?: Account;
-  proxyType: ProxyType;
-  delegate: Address;
   description: string;
   transaction: Transaction;
+  spawner: AccountId;
+  proxyType: ProxyType;
+  chain?: Chain;
+  account?: Account;
   proxiedAccount?: ProxiedAccount;
-
-  proxyDeposit: string;
-  proxyNumber: number;
 };
 
 const formInitiated = createEvent<Input>();
 const formSubmitted = createEvent();
 
-const $confirmStore = restore<Input>(formInitiated, null).reset(formSubmitted);
+const $confirmStore = restore<Input>(formInitiated, null);
 
 const $api = combine(
   {
@@ -41,9 +37,37 @@ const $initiatorWallet = combine(
     wallets: walletModel.$wallets,
   },
   ({ store, wallets }) => {
-    if (!store) return null;
+    if (!store || !store.account) return null;
 
     return walletUtils.getWalletById(wallets, store.account.walletId);
+  },
+  { skipVoid: false },
+);
+
+const $proxyWallet = combine(
+  {
+    store: $confirmStore,
+    wallets: walletModel.$wallets,
+  },
+  ({ store, wallets }) => {
+    if (!store || !store.account) return undefined;
+
+    return walletUtils.getWalletFilteredAccounts(wallets, {
+      accountFn: (a) => a.accountId === (store.account as ProxiedAccount).proxyAccountId,
+    });
+  },
+  { skipVoid: false },
+);
+
+const $signerWallet = combine(
+  {
+    store: $confirmStore,
+    wallets: walletModel.$wallets,
+  },
+  ({ store, wallets }) => {
+    if (!store || !store.account) return null;
+
+    return walletUtils.getWalletById(wallets, store.signatory?.walletId || store.account.walletId);
   },
   { skipVoid: false },
 );
@@ -61,29 +85,18 @@ const $proxiedWallet = combine(
   { skipVoid: false },
 );
 
-const $signerWallet = combine(
-  {
-    store: $confirmStore,
-    wallets: walletModel.$wallets,
-  },
-  ({ store, wallets }) => {
-    if (!store) return null;
-
-    return walletUtils.getWalletById(wallets, store.signatory?.walletId || store.account.walletId);
-  },
-  { skipVoid: false },
-);
-
 export const confirmModel = {
   $confirmStore,
   $initiatorWallet,
   $signerWallet,
+  $proxyWallet,
   $proxiedWallet,
   $api,
 
   events: {
     formInitiated,
   },
+
   output: {
     formSubmitted,
   },
